Fix dtype mismatch for GPU (bfloat16 support)

+++ papers/topological-coherence/experiments/gpu_hyperparameter_search.py
@@ -114,8 +114,11 @@ def apply_topology_constraint(model, topology, radius: float, alpha: float):
                     combined = topo_bias + causal_bias
                     new_mask = combined.unsqueeze(0).unsqueeze(0)
 
+                    # Cast to model dtype (float16/bfloat16)
+                    new_mask = new_mask.to(hidden_states.dtype)
+
                     if attention_mask is not None:
-                        attention_mask = attention_mask + new_mask.to(attention_mask.dtype)
+                        attention_mask = attention_mask + new_mask
                     else:
                         attention_mask = new_mask
